+++ AntrazShop/Areas/Admin/Views/Account/Login.cshtml
@@ -61,7 +61,8 @@
                             </fieldset>
                             <div class="flex justify-between items-center">
                                 <div class="flex gap10">
-                                   
+                                    <input class="" type="checkbox" id="signed">
+                                    <label class="body-text" for="signed">Giữ thông tin đăng nhập</label>
                                 </div>
                                 <a href="/admin/account/fogotpassword" class="body-text tf-color">Quên mật khẩu?</a>
                             </div>
