+++ AntrazShop/Areas/Admin/Views/Product/_CreateEditPartial.cshtml
@@ -32,7 +32,16 @@
                 <label for="brandid" class="body-title mb-10">Thương hiệu </label>
                 <div class="select">
                     <select id="brandid" class="mb-10 validate-input">
-        
+                        <option value="">Chọn thương hiệu</option>
+                        <option value="1">Apple</option>
+                        <option value="2">Samsung</option>
+                        <option value="3">Dell</option>
+                        <option value="4">Sony</option>
+                        <option value="5">Xiaomi</option>
+                        <option value="6">Acer</option>
+                        <option value="7">Lenovo</option>
+                        <option value="8">Asus</option>
+                        <option value="9">Logitech</option>
                     </select>
 
                     <div class="text-tiny error-message text-danger" style="font-style : italic"></div>
