+++ AntrazShop/Areas/Admin/Views/Shared/_Navbar.cshtml
@@ -150,12 +150,12 @@
 
                     <ul class="dropdown-menu dropdown-menu-end has-content" aria-labelledby="dropdownMenuButton1">
                         <li>
-                            <h6>Tin nhắn</h6>
+                            <h6>Message</h6>
                         </li>
-                @*         <li>
+                        <li>
                             <div class="noti-item w-full wg-user active">
                                 <div class="image">
-                                    <img src="images/avatar/user-11.png" alt="">
+                                    <img @* src="images/avatar/user-11.png"  *@alt="">
                                 </div>
                                 <div class="flex-grow">
                                     <div class="flex items-center justify-between">
@@ -165,9 +165,8 @@
                                     <div class="text-tiny">Hello?</div>
                                 </div>
                             </div>
-                        </li> *@
-
-                        <li><a href="#" class="tf-button w-full">Xem tất cả</a></li>
+                        </li>
+                        <li><a href="#" class="tf-button w-full">View all</a></li>
                     </ul>
                 </div>
             </div>
@@ -183,9 +182,9 @@
 
                     <ul class="dropdown-menu dropdown-menu-end has-content" aria-labelledby="dropdownMenuButton2">
                         <li>
-                            <h6>Thông báo</h6>
+                            <h6>Notifications</h6>
                         </li>
-                       @*  <li>
+                        <li>
                             <div class="message-item item-1">
                                 <div class="image">
                                     <i class="icon-noti-1"></i>
@@ -195,7 +194,9 @@
                                     <div class="text-tiny">Morbi sapien massa, ultricies at rhoncus at, ullamcorper nec diam</div>
                                 </div>
                             </div>
-                        </li> *@
+                        </li>
+
+                        <li><a href="#" class="tf-button w-full">View all</a></li>
                     </ul>
                 </div>
             </div>
@@ -229,7 +230,7 @@
                                 <div class="icon">
                                     <i class="icon-mail"></i>
                                 </div>
-                                <div class="body-title-2">Hòm thư</div>
+                                <div class="body-title-2">Mail</div>
                             </a>
                         </li>
                         <li>
