+++ AntrazShop/Controllers/API/ProductController.cs
@@ -144,17 +144,5 @@ namespace AntrazShop.Controllers.API
 			}
 			else return Ok("Xoá thành công!");
 		}
-
-		[Authorize(Policy = "CanReviewProduct")]
-		[HttpPost("Review")]
-		public async Task<IActionResult> AddReview(ReviewDTO dto)
-		{
-			var response = await _productService.AddReview(dto);
-			if (!response.IsSuccess)
-			{
-				return BadRequest(new { errors = response.Errors });
-			}
-			else return Ok(response.Data);
-		}
 	}
 }
