+++ AntrazShop/Helper/AutoMapperProfile.cs
@@ -37,8 +37,7 @@ namespace AntrazShop.Helper
 
 			// Review -> ProductReviewVM
 			CreateMap<Review, ProductReviewVM>()
-				.ForMember(dest => dest.UserName, opt => opt.MapFrom(src => src.User.Name))
-				.ForMember(dest => dest.Avatar, opt => opt.MapFrom(src => src.User.Avatar));
+				.ForMember(dest => dest.UserName, opt => opt.MapFrom(src => src.User.Name));
 
 			// ColorCapacity -> ProductColorCapacityVM
 			CreateMap<ColorCapacity, ProductColorCapacityVM>()
