+++ AntrazShop/Interfaces/Services/IShopService.cs
@@ -1,4 +1,3 @@
-using AntrazShop.Data;
 using AntrazShop.Helper;
 using AntrazShop.Models.DTOModels;
 using AntrazShop.Models.ViewModels;
@@ -9,8 +8,5 @@ namespace AntrazShop.Interfaces.Services
 	{
 		Task<ServiceResponse<(IEnumerable<ProductVM>, Paginate)>> GetProducts(ProductFilter filter, int currentPg,
 			int size);
-		Task<ServiceResponse<IEnumerable<ProductVM>>> GetProductsTopSale();
-		Task<ServiceResponse<IEnumerable<ProductVM>>> GetProductsTopSell();
-		Task<ServiceResponse<IEnumerable<ProductVM>>> GetProductsTopRating();
 	}
 }
