--- AntrazShop/Models/DTOModels/ReviewDTO.cs
@@ -1,10 +1,0 @@
-namespace AntrazShop.Models.DTOModels
-{
-	public class ReviewDTO
-	{
-		public int UserId { get; set; }
-		public int ColorCapacityId { get; set; }
-		public string Description { get; set; }
-		public int Rating { get; set; }
-	}
-}
