--- AntrazShop/Models/ViewModels/OrderOverviewVM.cs
@@ -1,10 +1,0 @@
-namespace AntrazShop.Models.ViewModels
-{
-	public class OrderOverviewVM
-	{
-		public List<int> NewOrdersCountDetails { get; set; }
-		public int ProcessOrderCount { get; set; }
-		public int FailedOrdersCount { get; set; }
-		public int CompletedOrderCount { get; set; }
-	}
-}
