+++ AntrazShop/Models/ViewModels/ProductReviewVM.cs
@@ -5,7 +5,6 @@ namespace AntrazShop.Models.ViewModels
 		public string UserName { get; set; }
 		public string Description { get; set; }
 		public double Rating { get; set; }
-		public string Avatar { get; set; }
 		public DateTime CreatedAt { get; set; }
 	}
 }
