+++ AntrazShop/Views/Account/Index.cshtml
@@ -71,11 +71,11 @@
                 <label class="body-title mb-20">Ảnh đại diện</label>
                     <div class="upload-image mb-16" style="margin-bottom:30px">
                         <div class="item" style="display:flex; justify-content:center">
-                            <label class="uploadfile mb-0" for="edit-avatar-input" style="gap:0; cursor:pointer;">
+                            <label class="uploadfile mb-0" for="edit-avatar-input" style="gap:0;">
                                 <!-- Ảnh -->
                                 <div class="img-preview-account" style="width:100%;">
                                     <img style="object-fit: contain; height:250px;" id="edit-avatar-preview">
-                                <input class="form-control-plaintext edit-information" disabled style="display:none" type="file" id="edit-avatar-input" accept="image/*">
+                                <input class="form-control-plaintext edit-information" style="display:none" type="file" id="edit-avatar-input" accept="image/*">
                                 </div>
                             </label>
                         </div>
