+++ AntrazShop/Views/Shop/_leftbar.cshtml
@@ -11,27 +11,6 @@
 
     <!--Price start-->
     <div class="tpshop__widget mb-30 pb-25">
-        <div>
-            <div class="form-check">
-                <input onclick="loadData()" class="form-check-input" value="null" type="radio" checked name="ascendingPrice" id="radioDefault1">
-                <label class="form-check-label" for="radioDefault1">
-                   Mới nhất
-                </label>
-            </div>
-            <div class="form-check">
-                <input onclick="loadData()" class="form-check-input" value="true" type="radio" name="ascendingPrice" id="radioDefault2">
-                <label class="form-check-label" for="radioDefault2">
-                    Giá tăng dần
-                </label>
-            </div>
-            <div class="form-check">
-                <input onclick="loadData()" class="form-check-input" value="false" type="radio" name="ascendingPrice" id="radioDefault3">
-                <label class="form-check-label" for="radioDefault3">
-                   Giá giảm dần
-                </label>
-            </div>
-        </div>
-        
         <h4 class="tpshop__widget-title mb-20">Mức giá</h4>
         <div class="productsidebar">
             <div class="productsidebar__head">
@@ -47,7 +26,6 @@
             <a onclick="filterPrice()" href="#">Lọc</a>
         </div>
     </div>
-
     <!--Price end-->
     <!--Sort by Price start-->
    @*  <div style="display:flex" class="tpshop__widget mb-30 pb-25">
