+++ AntrazShop/appsettings.json
@@ -1,14 +1,14 @@
 {
-  //"ConnectionStrings": {
-  //  "AntrazShop": "Data Source=LAPTOP-V249RLLF\\SQLEXPRESS;Initial Catalog=AntrazShopDb;Integrated Security=True;Trust Server Certificate=True"
-  //},
-
   "ConnectionStrings": {
-    "AntrazShop": "Data Source=remote-db7413.maychudns.net,1443;Initial Catalog=antraz_store_db3;Persist Security Info=True;User ID=lucdt123;Password=TranLuc0104;Trust Server Certificate=True"
+    "AntrazShop": "Data Source=LAPTOP-V249RLLF\\SQLEXPRESS;Initial Catalog=AntrazShopDb;Integrated Security=True;Trust Server Certificate=True"
   },
 
   //"ConnectionStrings": {
-  //  "AntrazShop": "Data Source=apple-rds.maychudns.net,1443;Initial Catalog=antraz_store_db3;Persist Security Info=True;User ID=lucdt123;Password=TranLuc0104;Trust Server Certificate=True"
+  //  "AntrazShop": "Data Source=remote-db7413.maychudns.net,1443;Initial Catalog=antraz_store_db2;Persist Security Info=True;User ID=lucdt456;Password=TranLuc0104;Trust Server Certificate=True"
+  //},
+
+  //"ConnectionStrings": {
+  //  "AntrazShop": "Data Source=apple-rds.maychudns.net,1443;Initial Catalog=antraz_store_db2;Persist Security Info=True;User ID=lucdt123;Password=TranLuc0104;Trust Server Certificate=True"
   //},
 
   "EmailSettings": {
@@ -25,7 +25,7 @@
     "Issuer": "https://localhost:7092/",
     "Audience": "https://localhost:7092/",
     "Key": "TranTheLuc2305NguyenThuyAn010402",
-    "TokenValidityMins": 300
+    "TokenValidityMins": 30
   },
 
   //"JwtConfig": {
