+++ AntrazShop/wwwroot/admin/js/api-jquery/product/edit.js
@@ -166,22 +166,8 @@ function loadEdit(id) {
 function LoadProductCC(productCCs) {
     $("#productCC-list").empty();
     $.each(productCCs, function (index, productCC) {
-        statusInt = productCC.status;
-        if (productCC.stock == 0) statusInt = 2;
-        let status;
-        switch (statusInt) {
-            case 0:
-                status = '<div class="block-pending">Ngừng bán</div>';
-                break;
-            case 1:
-                status = '<div class="block-available">Đang bán</div>';
-                break;
-            case 2:
-                status = '<div class="block-not-available">Hết hàng</div>';
-                break;
-            default:
-                status = "---";
-        }
+
+        let status = (productCC.status == 1) ? '<div class="block-available">Đang bán</div>' : '<div class="block-pending">Ngừng bán</div>';
 
         $("#productCC-list").append(
             ` <tr class="antraz-table-list">
@@ -290,15 +276,15 @@ function AddProductCC() {
                     contentType: false,
 
                     success: function (response) {
-                        Swal.fire({
-                            position: "center",
+                        swal.fire({
+                            title: "Thành công!",
                             icon: "success",
-                            title: "Thêm thành công!",
-                            showConfirmButton: false,
-                            timer: 1000
+                            draggable: true
+                        }).then(() => {
+                            loadEdit(idProduct);
+                            $('#productCCModal').modal('hide');
+
                         });
-                        loadEdit(idProduct);
-                        $('#productCCModal').modal('hide');
 
                     },
                     error: function (xhr, status, error) {
@@ -427,17 +413,17 @@ function SaveProductCC() {
                     contentType: false,
 
                     success: function (response) {
-                        Swal.fire({
-                            position: "center",
+                        swal.fire({
+                            title: "Thành công!",
                             icon: "success",
-                            title: "Cập nhật thành công!",
-                            showConfirmButton: false,
-                            timer: 1000
+                            draggable: true
+                        }).then(() => {
+                            let urlParams = new URLSearchParams(window.location.search);
+                            let id = urlParams.get('id');
+                            loadEdit(id);
+                            $('#productCCModal').modal('hide');
+
                         });
-                        let urlParams = new URLSearchParams(window.location.search);
-                        let id = urlParams.get('id');
-                        loadEdit(id);
-                        $('#productCCModal').modal('hide');
 
                     },
                     error: function (xhr, status, error) {
@@ -491,12 +477,10 @@ function DeleteProductCC(button) {
                     'Authorization': 'Bearer ' + token
                 },
                 success: function (response) {
-                    Swal.fire({
-                        position: "center",
-                        icon: "success",
-                        title: "Xoá thành công!",
-                        showConfirmButton: false,
-                        timer: 1000
+                    swalWithBootstrapButtons.fire({
+                        title: "Deleted!",
+                        text: "Your file has been deleted.",
+                        icon: "success"
                     });
                     let urlParams = new URLSearchParams(window.location.search);
                     let id = urlParams.get('id');
