+++ AntrazShop/wwwroot/admin/js/api-jquery/product/listproduct.js
@@ -179,7 +179,7 @@ function loadPageTableProduct(products) {
                 statusText = '<div class="block-not-available">Hết hàng</div>';
                 break;
             default:
-                statusText = "---";
+                status = "---";
         }
 
         // xử lý tên quá dài
