+++ AntrazShop/wwwroot/js/API/Shop/data-product-shop.js
@@ -8,19 +8,15 @@ var ProductList = [];
 var SwitchListView = 2;
 
 function loadData() {
-    let ascendingPrice = $('input[name="ascendingPrice"]:checked').val();
-    if (ascendingPrice == 'null') ascendingPrice = null;
-    if (ascendingPrice == 'true') ascendingPrice = true;
-    if (ascendingPrice == 'false') ascendingPrice = false;
-    console.log(ascendingPrice);
     let Filter = {
         brandIds: FListBrandIds,
         categoryIds: FListCategoryIds,
         searchText: FSearchText,
         minPrice: FPrice.Min,
         maxPrice: FPrice.Max,
-        ascendingPrice: ascendingPrice
+        ascendingPrice: null
     };
+
     window.scrollTo(0, 0);
 
     pager.numberShowing = $("#numberShowing").val();
@@ -72,7 +68,7 @@ function loadData() {
             }
         },
         error: function (xhr, status, error) {
-            handleAjaxError(xhr, status, error, "Lỗi khi load sản phẩm");
+            console.error("Lỗi: ", error);
         }
     });
 }
@@ -125,8 +121,8 @@ function loadProducts2() {
             }
         }
 
-        let min_price_sale = (product.minPrice - product.discountAmount).toLocaleString('vi-VN') + ' VNĐ';
-        let min_price = product.minPrice.toLocaleString('vi-VN') + ' VNĐ';
+        let min_price_sale = (product.minPrice - product.discountAmount).toLocaleString('vi-VN') + ' đ';
+        let min_price = product.minPrice.toLocaleString('vi-VN') + ' đ';
 
         $("#list-product-2").append(`
             <div class="col">
@@ -164,11 +160,9 @@ function loadProducts2() {
                             ${starRate}
                         </div>
                         <div class="tpproduct__price">
-                   <span>${min_price_sale}</span>
-               </div>
-                <div class="tpproduct__price">
-                   <del>${min_price}</del>
-               </div>
+                            <span>${min_price_sale}</span>
+                            <del>${min_price}</del>
+                        </div>
                     </div>
                     <div class="tpproduct__hover-text">
                         <div class="tpproduct__hover-btn d-flex justify-content-center mb-10">
@@ -218,16 +212,17 @@ function loadProducts1() {
             }
         }
 
-        let mbClass = (index < 4) ? "mb-20" : "";
+        let mbClass = (index < 4) ? "mb-20" : ""; 
 
-        let min_price_sale = (product.minPrice - product.discountAmount).toLocaleString('vi-VN') + ' VNĐ';
-        let min_price = product.minPrice.toLocaleString('vi-VN') + ' VNĐ';
+        let min_price_sale = (product.minPrice - product.discountAmount).toLocaleString('vi-VN') + ' đ';
+        let min_price = product.minPrice.toLocaleString('vi-VN') + ' đ';
 
         $("#list-product-1").append(`
-            <div class="col">
-                <div class="tpproduct p-relative ${mbClass}">
+            <div class="col target-product">
+             <input type="hidden" class="hidden-id" value="${product.id}" />
+                <div class="tpproduct p-relative ${mbClass}" style="z-index: ${6 - index};">
                     <div class="tpproduct__thumb p-relative text-center">
-                        <a href="/shop/productdetail?id=${product.id}">
+                        <a href="#">
                             <img src="/admin/imgs/product/${product.folderImage}/${product.imageView}" alt="">
                         </a>
                         <a class="tpproduct__thumb-img" href="/shop/productdetail?id=${product.id}">
@@ -238,7 +233,7 @@ function loadProducts1() {
                             <span class="tpproduct__info-hot bage__hot">HOT</span>
                         </div>
                         <div class="tpproduct__shopping">
-                            <a class="tpproduct__shopping-wishlist">
+                            <a class="tpproduct__shopping-wishlist" href="wishlist.html">
                                 <i class="icon-heart icons"></i>
                             </a>
                             <a class="tpproduct__shopping-cart" href="/shop/productdetail?id=${product.id}">
@@ -262,14 +257,14 @@ function loadProducts1() {
                             <del>${min_price}</del>
                         </div>
                     </div>
-                    <div class="tpproduct__hover-text">
+                    <div class="tpproduct__hover-text" style="z-index: 999;">
                         <div class="tpproduct__hover-btn d-flex justify-content-center mb-10">
-                            <button class="tp-btn-2" onclick="addToCart(${product.productCCs[0].id})">Thêm vào giỏ</button>
+                            <button class="tp-btn-2"  onclick="addToCart(${product.productCCs[0].id})">Thêm vào giỏ</button>
                         </div>
                         <div class="tpproduct__descrip">
                             <ul>
                                 <li>Đã bán ${product.soldAmount}</li>
-                                <li>Tồn kho: ${product.totalStock}</li>
+                                <li>Tồn kho: ${product.totalStock} </li>
                             </ul>
                         </div>
                     </div>
@@ -278,6 +273,7 @@ function loadProducts1() {
         `);
     });
 }
+
 
 function loadProducts3() {
     $("#list-product-3").empty();
@@ -310,8 +306,8 @@ function loadProducts3() {
             }
         }
 
-        let min_price_sale = (product.minPrice - product.discountAmount).toLocaleString('vi-VN') + ' VNĐ';
-        let min_price = product.minPrice.toLocaleString('vi-VN') + ' VNĐ';
+        let min_price_sale = (product.minPrice - product.discountAmount).toLocaleString('vi-VN') + ' đ';
+        let min_price = product.minPrice.toLocaleString('vi-VN') + ' đ';
 
         $("#list-product-3").append(`
             <div class="col-lg-12" target-product>
@@ -345,11 +341,8 @@ function loadProducts3() {
                         <h4 class="tplist__instock">
                             Tồn kho: <span>${product.totalStock}</span>
                         </h4>
-                        <h3 class="tplist__count">${min_price_sale}</h3>
-                          <div class="tpproduct__price">
-                   <del>${min_price}</del>
-               </div>
-                        <button class="tp-btn-2 mt-10 mb-10" onclick="addToCart(${product.productCCs[0].id})">Thêm vào giỏ</button>
+                        <h3 class="tplist__count mb-15">${min_price_sale}</del></h3>
+                        <button class="tp-btn-2 mb-10" onclick="addToCart(${product.productCCs[0].id})">Thêm vào giỏ</button>
                         <div class="tplist__shopping">
                             <a href="#"><i class="icon-heart icons"></i> wishlist</a>
                         </div>
